+++ braindecode/preprocessing/__init__.py
@@ -1,4 +1,4 @@
-from .preprocess import (zscore, scale, exponential_moving_demean,
+from .preprocess import (scale, exponential_moving_demean,
                          exponential_moving_standardize, filterbank,
                          preprocess, Preprocessor)
 from .windowers import (create_windows_from_events, create_fixed_length_windows,